frontend: extract remaining route page implementations into features
