+++ frontend/src/features/auth/pages/SignUpPage.tsx
@@ -1,0 +1,14 @@
+import { SignUp } from "@clerk/react"
+import type { ReactElement } from "react"
+import { AuthPageShell } from "../../../components/app/AuthPageShell.tsx"
+
+export function SignUpPage({ redirect }: { redirect: string }): ReactElement {
+  return (
+    <AuthPageShell
+      title="Start your empire"
+      description="Create your account to join a lobby, claim a faction, and start competing for long-term control of the galaxy."
+    >
+      <SignUp signInUrl={`/sign-in?redirect=${redirect}`} fallbackRedirectUrl={redirect} />
+    </AuthPageShell>
+  )
+}
