+++ frontend/src/routes/_app.games.$gameId.tsx
@@ -1,28 +1,14 @@
-import { useMutation, useQuery, type UseMutationOptions } from "@tanstack/react-query"
-import { createFileRoute, Navigate, redirect, useNavigate } from "@tanstack/react-router"
-import type { ReactElement } from "react"
+import { createFileRoute, redirect } from "@tanstack/react-router"
 import { z } from "zod"
-import type * as ApiTypes from "@api-types"
-import { GameStatusBadge } from "../components/play/GameStatusBadge.tsx"
-import { PageHeader } from "../components/app/PageHeader.tsx"
-import { Button } from "../components/ui/button.tsx"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "../components/ui/card.tsx"
-import { Separator } from "../components/ui/separator.tsx"
-import { Skeleton } from "../components/ui/skeleton.tsx"
-import { useBackendApiClient } from "../contexts/BackendApiClientContext.tsx"
-import { useLogger } from "../contexts/LoggerContext.tsx"
-import { formatGameSummaryStatus } from "../lib/formatGameSummaryStatus.ts"
-import { timeAgo } from "../lib/timeAgo.ts"
+import { GameLobbyPage } from "../features/games/pages/GameLobbyPage.tsx"
 
 const paramsSchema = z.object({
   gameId: z.coerce.number(),
 })
 
 export const Route = createFileRoute("/_app/games/$gameId")({
-  component: GameLobby,
-  params: {
-    parse: (params) => paramsSchema.parse(params),
-  },
+  component: GameLobbyRoute,
+  params: { parse: (params) => paramsSchema.parse(params) },
   onError: (error) => {
     if (error?.routerCode === "PARSE_PARAMS") {
       // eslint-disable-next-line @typescript-eslint/only-throw-error -- That's how tanstack works
@@ -31,172 +17,7 @@ export const Route = createFileRoute("/_app/games/$gameId")({
   },
 })
 
-function GameLobby(): ReactElement {
-  const logger = useLogger()
+function GameLobbyRoute() {
   const { gameId } = Route.useParams()
-  const backendApiClient = useBackendApiClient()
-  const gameQuery = useQuery(backendApiClient.games.getSummaryById.queryOptions({ gameId }))
-
-  if (gameQuery.isPending) {
-    return <GameLobbyLoadingState />
-  }
-
-  if (gameQuery.isError) {
-    logger.error("Could not fetch game", { gameId, error: gameQuery.error.message })
-    return <Navigate to="/games" />
-  }
-
-  const gameData = gameQuery.data
-
-  return (
-    <div className="mx-auto flex w-full max-w-5xl flex-col gap-6">
-      <Game game={gameData.game} />
-    </div>
-  )
-}
-
-function Game({ game }: { game: ApiTypes.GameSummary }): ReactElement {
-  const navigate = useNavigate()
-  const backendApiClient = useBackendApiClient()
-  const joinGame = useMutation(backendApiClient.games.join.mutationOptions() as UseMutationOptions<unknown, Error, { gameId: number }>)
-  const leaveGame = useMutation(backendApiClient.games.leave.mutationOptions() as UseMutationOptions<unknown, Error, { gameId: number }>)
-  const startGame = useMutation(backendApiClient.games.start.mutationOptions() as UseMutationOptions<unknown, Error, { gameId: number }>)
-
-  return (
-    <div className="flex flex-col gap-6">
-      <PageHeader
-        title={game.name}
-        description={`Game #${game.id} created ${timeAgo(game.createdAt)}.`}
-        actions={<GameStatusBadge status={game.status} />}
-      />
-      <Card className="border border-border/60">
-        <CardHeader>
-          <CardTitle>Lobby details</CardTitle>
-          <CardDescription>Review the roster, join or leave the lobby, and start once the game is ready.</CardDescription>
-        </CardHeader>
-        <CardContent className="flex flex-col gap-6">
-          <div className="grid gap-4 md:grid-cols-2">
-            <DetailBlock label="Status" value={formatGameSummaryStatus(game.status)} />
-            <DetailBlock label="Seats" value={`${game.players.length}/${game.nbSeats} players`} />
-            <div className="space-y-1">
-              <div className="text-xs font-medium tracking-[0.2em] text-muted-foreground uppercase">Creator</div>
-              <Player player={game.creator} />
-            </div>
-            <DetailBlock label="Created" value={timeAgo(game.createdAt)} />
-            {game.winnerPlayerId !== null ? <DetailBlock label="Winner" value={getWinnerLabel(game)} /> : null}
-          </div>
-          <Separator />
-          <div className="space-y-3">
-            <div className="text-xs font-medium tracking-[0.2em] text-muted-foreground uppercase">
-              Players ({game.players.length}/{game.nbSeats})
-            </div>
-            <div className="grid gap-2">
-              {game.players.map((player) => (
-                <Player player={player} key={player.id} />
-              ))}
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-      <div className="flex flex-wrap gap-3">
-        {game.canJoin && (
-          <Button
-            disabled={joinGame.isPending}
-            onClick={() => {
-              joinGame.mutate({ gameId: game.id })
-            }}
-          >
-            Join game
-          </Button>
-        )}
-        {game.canLeave && (
-          <Button
-            variant="outline"
-            disabled={leaveGame.isPending}
-            onClick={() => {
-              leaveGame.mutate({ gameId: game.id })
-            }}
-          >
-            Leave game
-          </Button>
-        )}
-        {game.canStart && (
-          <Button
-            disabled={startGame.isPending}
-            onClick={() => {
-              startGame.mutate({ gameId: game.id })
-            }}
-          >
-            Start game
-          </Button>
-        )}
-        {game.status === "STARTED" && (
-          <Button
-            variant="secondary"
-            disabled={startGame.isPending}
-            onClick={() => {
-              void navigate({ to: "/play/$gameId", params: { gameId: game.id } })
-            }}
-          >
-            Open game
-          </Button>
-        )}
-      </div>
-    </div>
-  )
-}
-
-function DetailBlock({ label, value }: { label: string; value: string }): ReactElement {
-  return (
-    <div className="space-y-1">
-      <div className="text-xs font-medium tracking-[0.2em] text-muted-foreground uppercase">{label}</div>
-      <div className="text-sm text-foreground">{value}</div>
-    </div>
-  )
-}
-
-function Player({ player }: { player: ApiTypes.GameSummaryPlayer }): ReactElement {
-  return (
-    <div className="rounded-3xl border border-border/60 bg-muted/20 px-4 py-3">
-      <div className="font-medium text-foreground">{player.alias ?? `Player ${player.id}`}</div>
-    </div>
-  )
-}
-
-function GameLobbyLoadingState(): ReactElement {
-  return (
-    <div className="mx-auto flex w-full max-w-5xl flex-col gap-6">
-      <PageHeader
-        title="Loading game"
-        description="Fetching lobby details and available actions."
-        actions={<Skeleton className="h-6 w-28 rounded-full" />}
-      />
-      <Card className="border border-border/60">
-        <CardHeader>
-          <Skeleton className="h-6 w-32" />
-          <Skeleton className="h-4 w-80 max-w-full" />
-        </CardHeader>
-        <CardContent className="flex flex-col gap-3">
-          <Skeleton className="h-20 w-full" />
-          <Skeleton className="h-px w-full" />
-          <Skeleton className="h-14 w-full" />
-          <Skeleton className="h-14 w-full" />
-        </CardContent>
-      </Card>
-      <div className="flex gap-3">
-        <Skeleton className="h-9 w-24 rounded-4xl" />
-        <Skeleton className="h-9 w-24 rounded-4xl" />
-      </div>
-    </div>
-  )
-}
-
-function getWinnerLabel(game: ApiTypes.GameSummary): string {
-  const winner = [game.creator, ...game.players].find((player) => player.id === game.winnerPlayerId)
-
-  if (winner === undefined) {
-    return `Player ${game.winnerPlayerId}`
-  }
-
-  return winner.alias ?? `Player ${winner.id}`
+  return <GameLobbyPage gameId={gameId} />
 }
