+++ frontend/src/routes/_app.sign-in.tsx
@@ -1,27 +1,17 @@
 import { createFileRoute } from "@tanstack/react-router"
-import type { ReactElement } from "react"
-import { SignIn } from "@clerk/react"
 import { z } from "zod"
-import { AuthPageShell } from "../components/app/AuthPageShell.tsx"
+import { SignInPage } from "../features/auth/pages/SignInPage.tsx"
 
 const RedirectSchema = z.object({
   redirect: z.string().default("/games"),
 })
 
 export const Route = createFileRoute("/_app/sign-in")({
-  component: SignInComponent,
+  component: SignInRoute,
   validateSearch: RedirectSchema,
 })
 
-function SignInComponent(): ReactElement {
+function SignInRoute() {
   const { redirect } = Route.useSearch()
-
-  return (
-    <AuthPageShell
-      title="Welcome back"
-      description="Sign in to manage your ongoing campaigns, review intel, and jump back into your empire before the next tick lands."
-    >
-      <SignIn signUpUrl={`/sign-up?redirect=${redirect}`} fallbackRedirectUrl={redirect} />
-    </AuthPageShell>
-  )
+  return <SignInPage redirect={redirect} />
 }
