+++ frontend/src/routes/play.$gameId.tsx
@@ -1,11 +1,6 @@
-import { useQuery } from "@tanstack/react-query"
-import { createFileRoute, Navigate, Outlet, redirect } from "@tanstack/react-router"
-import type { ReactElement } from "react"
+import { createFileRoute, redirect } from "@tanstack/react-router"
 import { z } from "zod"
-import { GameLayout, GameLayoutSkeleton } from "../components/play/GameLayout.tsx"
-import { useBackendApiClient } from "../contexts/BackendApiClientContext.tsx"
-import { useLogger } from "../contexts/LoggerContext.tsx"
-import { PlayGameContextProvider, type PlayGameContextValue } from "../contexts/PlayGameContext.tsx"
+import { PlayGameLayoutPage } from "../features/play/pages/PlayGameLayoutPage.tsx"
 import { privateRoute } from "../privateRoute.ts"
 
 const paramsSchema = z.object({
@@ -13,11 +8,9 @@ const paramsSchema = z.object({
 })
 
 export const Route = createFileRoute("/play/$gameId")({
-  component: PlayGameLayout,
+  component: PlayGameLayoutRoute,
   beforeLoad: privateRoute,
-  params: {
-    parse: (params) => paramsSchema.parse(params),
-  },
+  params: { parse: (params) => paramsSchema.parse(params) },
   onError: (error) => {
     if (error?.routerCode === "PARSE_PARAMS") {
       // eslint-disable-next-line @typescript-eslint/only-throw-error -- That's how tanstack works
@@ -26,37 +19,7 @@ export const Route = createFileRoute("/play/$gameId")({
   },
 })
 
-function PlayGameLayout(): ReactElement {
-  const logger = useLogger()
+function PlayGameLayoutRoute() {
   const { gameId } = Route.useParams()
-  const backendApiClient = useBackendApiClient()
-  const gameQuery = useQuery(backendApiClient.games.getSummaryById.queryOptions({ gameId }))
-  const gameStateQuery = useQuery(backendApiClient.gameStates.getById.queryOptions({ gameId }))
-
-  if (gameQuery.isPending || gameStateQuery.isPending) {
-    return <GameLayoutSkeleton />
-  }
-
-  if (gameQuery.isError) {
-    logger.error("Could not fetch game", { gameId, error: gameQuery.error.message })
-    return <Navigate to="/games" />
-  }
-
-  if (gameStateQuery.isError) {
-    logger.error("Could not fetch game state", { gameId, error: gameStateQuery.error.message })
-    return <Navigate to="/games" />
-  }
-
-  const context: PlayGameContextValue = {
-    game: gameQuery.data.game,
-    gameState: gameStateQuery.data.gameState,
-  }
-
-  return (
-    <PlayGameContextProvider value={context}>
-      <GameLayout game={context.game} gameState={context.gameState}>
-        <Outlet />
-      </GameLayout>
-    </PlayGameContextProvider>
-  )
+  return <PlayGameLayoutPage gameId={gameId} />
 }
